+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Yourloops data visualization for diabetes device data
 Library for Tidepool's timeline-style diabetes data visualization(s) used in Blip and was forked from tidepool/tideline.
 
+## unreleased
+### Added
+- PT-865 Display physical activity in Bolus/Food section
+
 ## 1.4.0 - 2019-12-04
 ### Added
 - PT-726 Review the logo/picto for insulin sites